Get Basilisk II working with the create-react-app harness
Includes a partial refactor of the UI thread code to be more type
safe and object oriented.

To get the Webpack/create-react-app dev server to set the expected
headers (that the Cloudflare Worker would also set), a custom middleware
is added. CRA does not explicitly allow this, but the custom proxy
support (https://create-react-app.dev/docs/proxying-api-requests-in-development/#configuring-the-proxy-manually)
gives us arbitrary access to the Express app intstance.

The emscripten-generated .js and .wasm files had to be renamed to
.jsz and .wasmz so that we could use file-loader with them, for
whatever reason using a `file-loader!` Webpack import syntax did
not work.