Add support for more fine-grained and efficient loading of software
Instead of a whole disk image that has to be loaded at startup, we
use lazily-loaded Emscripten files in a "Library" directory.

Each app is packaged as a .zip and a .json manifest, there's helper
`extractDirectory` function in the worker that can be manually invoked
to generate them.

Appears to work, with a few caveats:
- Icons for some applications (e.g. Marathon, but not ResEdit) are not
  set correctly. Rebuilding the desktop file appears to fix this, but
  that would involve touching every file, which negates the benefits
  of the lazy loading.
- The lazy loading when serving in production does not support range
  requests, and thus the Emscripten JS support library ends up loading
  all of the file contents as soon as the folder is opened, which leads
  to UI freezes.