Make lazily-loaded applications have correct icons
We were generating files with a kHasBeenInited flag, thus when the
Finder encountered them it did not register the application and its
document types based on BNDL resources. Clear that flag when generating
the archive zip so that the Finder will re-register the app. Required
further tweaking to get location information to be preserved (thanks
1995 Apple for having good documentation).

To make this easier to debug, add a mode where the FInfo struct (stored
in .finf/ directories) can be visually inspected in the dev server.

Also add a versioning scheme to library archives so that we can cache
bust when updating the definitions (except that there's a bit of catch
22 where it's stored in the manifest JSON, which is also served behind
a URL -- for now add a cache busting parameter there too to always fetch
the manifest, but we should fix that).

Followup to 7ea4f6e6bf762ec59f53ff76b239d5adb0ae3cbd