Switch handling of library zip files to being done in a service worker
Fetching the zip file can take a while (and blocks the emulator thread
while it happens), this way we can know when it happens and have the
option of displaying progress (though that is not implemented yet).

Also has the benefit of making the local/dev and prod code paths the same.