Switch to our version of lazy files
The one that was built in to Emscripten was proving too finicky. Besides
the previously uncovered issue with requiring HEAD requests (which don't
work for file uploads), it was also preventing Glider from working. That
program appears to modify some of its own files and truncate them, but
LazyUint8Array as implemented by Emscripten does not have subarray,
which ftruncate -> FS.ftruncate -> FS.truncate -> MEMFS.node_ops.setattr
-> resizeFileStorage ends up relying on.

Switch to a much simpler lazy file implementation that creates a real
Uint8Array and lets Emcripten/MEMFS own it, and the hooks the `read`
operation to populate it on demand.

Main change is that the size of the lazy file needs to be known ahead
of time, but we could even work around that if we wanted to dynamically
replace the `contents` property of the node.