Add importer for Internet Archive/Emularity disk images
Uses https://github.com/elliotnunn/machfs to parse the disk images
and convert them into the same kind of zip files that the extractor
generates. machfs did not preserve folder DInfo and DXInfo structs,
a fork with mihaip/machfs@5992e174ed569f7929985cf35d4467b2b1e093f1 is
used to have that ability.

The fixup of FInfo/DInfo flags and locations was moved from write
time to read time (in the service worker), to avoid having to repeat
the logic in the image extractor.