Switch to loading the Macintosh HD disk image on-demand in chunks
Unlike our lazy files (createLazyFile from emulator-worker-lazy-file),
this loads separate parts of the file in chunks, instead of the entire
file.

Unlike the Emscripten FS.createLazyFile API, this supports compression
(because we split the file into individually Brotli-compressed chunks,
instead of using HTTP range requests). It also supports prefetching,
so that we can preload the sets of chunks that are needed for the
Mac startup. This helps to speed things up, since normal chunk loading
is synchronous, thus we end up blocking the emulator frequently if we
have to do things on-demand.

To get prefetching working and to get more control, the actual chunk
fetching is done in a service worker (and cached using the worker cache
API). The emulator worker still thinks it's doing synchronous
XMLHttpRequests, but they're actually intercepted by the service worker.

This helps a little bit with the uncached time-to-first-idlewait (from
~3.3 seconds to ~2.5 seconds) since we're no longer loading parts of the
disk image that are not needed. However, the bigger benefit is that this
unblocks having a much larger disk image, instead of using ExtFS/the
Emscripten file system for the software library. ExtFS and/or the
File System Manager interface that it builds on don't appear to provide
the level of compatibility that old software expects (e.g. Indiana Jones
and the Last Crusade and Battle Chess refuse to launch because they can't
find certain files), so the goal is to instead have that software
directly installed in the disk image, while still downloading it
on-demand.