Persist some data in the simulator across runs.
The contents of the "Saved" folder that's mounted in the Mac via ExtFS
are saved in IndexedDB. This us done using a homegrown method that
extracts the contents during emulator unloading and uses the idb-keyval
wrapper to store to. Emscripten's IDBFS is not an options since it
relies on an event loop, which not an option.

Communication of the emulator stop is done via a synthetic "stop"
event so that we don't need to set up an additional communication path
for it.