Catch exceptions when handling file uploads
It may fail (due to the Emscripten filesytem rejecting it if a file
already exists or the like), better to keep going than to pause.

Do this at a higher level (in the idlewait iteration) because there
may be other causes of exception, and we want to ensure that
Basilisk's emulation loop keeps going.