Handle writes in chunked files before reads.
We would not mark the chunks that the writes intersected with as loaded,
and therefore if there was a subsequent read we would load the chunk
and clobber what was written.

Would manifest itself when running an installer -- the written files
would usually end up being empty because they corresponded to empty
chunks.