Add basic support for dragging in CD-ROM images
They're written to the filesystem and the Basilisk II preferences file
is updated with extra `cdrom` lines to pick them up.

This requires restarting the entire simulator, which is easiest to do
by stopping and restarting the entire worker. This means that filesystem
changes outside of the "Saved" folder are not persisted, and is a bit
clunky. We could perhaps modify the cdrom.cpp handling in Basilisk II
to not require the files to be present at startup, and trigger
CDROMMountVolume() calls when they are actually created, but that is
a more invasive change.

Fixes #26