Rebuild Infinite HD Desktop DB as part of the image creation process
machfs only writes a stubbed out Desktop DB, and the format has not been
documented. Switch instead to not writing a Desktop DB at all (and not
preserving one from the base image). To get the Desktop DB built, we
launch the (native Mac) version of Basilisk II pointed at built image,
which will cause the Finder to rebuild the DB at startup. To make the
process automated, the disk image that we use has an AppleScript as
a startup item that runs the following code, so that it shuts down
once the DB has been rebuilt (and the disk is mounted):

```
on idle
  if (list disks) contains "Infinite HD" then
    tell application "Finder" to shut down
  end if
  return 1
end idle
```

(the return 1 from the idle handler makes it check every second, otherwise
the default is every 30 seconds)

This appears to work end-to-end, except there are still some issues
with the B*-tree data structures generated by machfs, and when populating
the image completely the disk traversal necessary for the Desktop DB
rebuild does not work correctly (it does work when only using a subset
of files).

For #4