Add debugging code for logging AppleTalk broadcast packets
Used to investigate why the machine pauses for ~5 seconds when AppleTalk
is enabled. This appears to be intentional, it sends out repeated AARP
probe requests with the tentative address that it has chosen, and wait
to see if it hears back about any other machine that has chosen it (see
Dynamic protocol address assignment in Chapter 2 of Inside AppleTalk).

There does not appear to be a way to short-circuit the AARP probing, so
this is left in as debugging code that's disabled by default (the intent
behind handleEthernetWrite was that some writes could be handled locally,
and a response be generated immediately).

For #53