Add Cloudflare Worker-based EthernetProvider
Uses durable objects to route ethernet packets between clients in a
zone. Nothing is actually persisted in storage, we use the in-memory
state of the durable object, since it only makes sense to broadcast
packets while to other clients that are currently actively connected.

There is also a debug /zone/<name>/list UI to list currently connected
clients.

For #53