Implement clipboard read and write support
For now only clipboard text is synchronized between the emulator and
the host. The Clipboard API is async, so we need to pre-emptively send
the contents to the emulator worker (via a SharedArrayBuffer) area of
time, so that by the time the clip_js.cpp GetScrap asks for it, it can
be returned immediately.

Fixes #64