Generate stickies during image building process.
Allows the same text to be used on all three disk images without having
to manually duplicate it. Paves the way for dynamically-generated content
(e.g. a changelog).

This involves running the boot disk images through machfs, which appears
to mostly work, except for Mac OS 8, which insists on rebuilding the
desktop database.