Reduce disk images
5be16f62c095759e1907d9a68301e6bf095bcf6b had a copy/paste error, where
we were using 1GB for the boot imags, even though they're usually
only 100MB. Preserve the size of the input image.

Reduce the size of the Infinite HD image by a bit (900MB instead of 1GB).
This makes it more likely to fit in more memory-constrained environments
(like 32-bit browsers and iOS 16). The correct fix is still to not
have the entire chunked file pre-allocate the Uint8Array (#56), but that
still requires further investigations.