Add Mac OS 9 disk image
Mac OS 9 is rather hard to fit into 100MB, but for now it does work.
Switching to Gif LFS may be a good idea.

Stickies layout had to be tweaked a bit because the large font is
larger and the "sister sites" sticky is now wider.

Also adds a DEBUG_SYSTEM_FILTER environment variable to allow a single
OS's disk image to be rebuilt, to speed up iteration.

Also includes associated changes to make macos9.app work.

Updates #34