Work around sound not working in Mac OS 8 and Mac OS 9
Appears to be due to corruption introduced by modifying the images with
machfs, which in turn results in the sound driver resource not being
loaded. I do not have a fix for that yet (tracked in #84), but as a
workaround we remove the need to read-modify-write the disk image with
machfs.

The main reason why we were doing that was to insert a dynamically-
generated stickies file (see 5be16f62c095759e1907d9a68301e6bf095bcf6b
and 6b047582a72dfd889910f7d3540f48a48b879c84). However, we can do that
with direct manipulation of the disk image, instead of doing HFS
operations. We do this by generating a placeholder "Stickies file" file
that we copy to the disk image by hand, and then replace it with the
actual file when building images (the images are defragmented with
Speed Disk from Norton Utilities to ensure the placeholder data is
continuous). The load-placeholder-stickies-file script helps with this.

Mac OS 9 still does not have sound by default, because the "Built-in"
source is not selected. It will likely require some tweaks to the PRAM
to fix that.

Updates #67