Handle mounting of additional disk images after the first restart
We were passing in the disk image's ArrayBuffer as a transferable
parameter to postMessage. This worked the first time, but on subsequent
restarts we the buffer had already been transferred, so an exception
would be thrown.

Redo how disk images are sent to the worker to instead use lazy files
(via blob URLs), so that we don't need to read (or transfer) the buffer
at all.

Also clear the screen as soon as the worker is terminated, so that we
don't show the last blitted frame so long longer (until the emulator
sends a new frame after initialization).

Fixes #77