Move disk and machine selection out of the Mac component
More generally separates out disk and machine selection from the domain,
so that we can start the emulator with different combinations (for now
only specified via query parameters).

This allows more bootable combinations, for example running System 7.5.3
on a Mac Plus or Mac II, or Mac OS 9.0.4 on an Old World PowerMac.

For #100