Add TeachText and Welcome placeholder to System 1.0 image
Also switch it to being a 400K disk image, instead of a ~256K one with
no empty room.

Files were copied from the System 6.0.8 image (by mounting both in the
Mac Plus/System 6.0.8 emulator, which can read from HFS and write to
MFS). I had to rebuild the desktop of the System 1.0 image after,
since it was ending up with a blank icon.

Exclude references to "The Outside World" for images that are meant to
be run in Mini vMac, since it doesn't have a similar external file
system mechanism.

Updates #100