Clean up ROM/machine names a bit
The Old World ROM is a 7200/7500/8500/9500 ROM. In the UI the machine
is called a 9500, since that matches what rom_patches.cpp injects
in SheepShaver.

The New World ROM is version 1.6, but since it's not associated with
a specific model, it's left with that name. In the UI it's labeled as
a G3, since that's the generation that most closely matches the
timeframe (early 1999).

Updates #100