Add System 7.5.2
Installed from https://archive.org/details/pm-9500-120-9500-132 (converted
from .mdf to .iso with mdf2iso).

Boots in the native build of SheepShaver, but not in the Emscripten
builds (fails with an "memory access out of bounds" exception). The
native build does not appear to trigger any SIGSEGVs, so this needs
further investigation.

Updates #100