Use a link for the demo video
Redoes f70f8a1fce3cdfecff0d8ba1c760591f8f54a4e7 to just link to the
video instead of embedding it. This is because YouTube embeds are still
not fully compatible with cross-origin isolation (see https://issuetracker.google.com/issues/240387105).

<iframe credentialless> should help, but that's not uniformly implemented
yet, so a link is easier. We link to the embed URL because it's less
likely to have pre-roll ads.

Updates #100