Ignore cache-related errors when fetching chunks in the service worker
Appearst to be the more common error when in Twitter's in-app browser
(the out-of-memory error from 89cb33397881391d5ca0f31552cdd19ae97d24f8
only happens when reloading). I'm guessing that SFSafariViewController
has a much smaller quota (or no quota at all?) and is not able to open
or write to the cache.

If we ignore errors then we're able to load successfully.