Ignore cache-related errors when fetching chunks in the service worker
Appearst to be the more common error when in Twitter's in-app browser
(the out-of-memory error from 6ec5d37a5c8036bea6bda18c0b752be4d6b3ad21
only happens when reloading). I'm guessing that SFSafariViewController
has a much smaller quota (or no quota at all?) and is not able to open
or write to the cache.

If we ignore errors then we're able to load successfully.