Allow disk images to be mounted without restarting Basilisk II or SheepShaver
Picks up mihaip/macemu@3b8fb53bf6c1ece90f316ab89ba6e4c25ae82d97,
which handles mountable media in the disk_js implementation.

A EmulatorCDROMDrive class is added which wraps an existing
EmulatorWorkerDisk when a disk is inserted but is also a
placeholder that can be polled when empty.

Fixes #49