Checkpoint CD-ROM Library support
CD-ROM disk images are specified via JSON manifests that include a source
URL and a cover image.

Those are processed by an import-cd-roms.py script into a single CD-ROMS.json
file for the UI to use. Besides combining the individual manifests, it also:
- Does a HEAD request (to get the overall size and ensure that range requests
  are supported).
- Fetches and resizes the cover image to a consistent 512x512 size.

On the UI side, there's a placeholder browser (enabled via a ?cdroms=true query
parameter) that shows a bottom-anchored pane with all of the CD-ROMs. Loading
reuses some of the file upload mechanism to send the manifest to the worker.

The worker then generates a chunked file spec (assuming 128K chunks) and mounts
the CD-ROM like any other disk image.

The chunked file is handled by a /CD-ROM/ handler (implemented both locally and
in the Cloudflare Worker). The handler takes a chunk file name (which indicates
the range that should be fetched) and does an HTTP range request to the source
URL for it.

Updates #166
Updates #173