Use dyanmic imports for the emulator and disk definitions
Initially the main JS payload was 716K (262K after gzip). By wrapping
the Mac component in React.lazy(), it dropped to 580K (219K after gzip).
By making the disk definitions lazy, it dropped to 181K (60K after gzip).

This does add another more round-trip for loading a specific emulator,
but it does not not appear to be a noticeable delay (it would be two,
but we prefetch the disk definition at the same time that we start to
render the component).