Handle space-encoded assets files in the worker URL
We run into cloudflare/kv-asset-handler#331 because we use a custom
mapRequestToAsset function. This was not an issue before because we
always used file names without spaces, but starting with the dynamic
imports for disk definitions in 9eb9fc6438332a73506563b76934fef413047afc
we end up with "Sytem 1.0-<hash>.dsk.js" and similar files.