Handle space-encoded assets files in the worker URL
We run into cloudflare/kv-asset-handler#331 because we use a custom
mapRequestToAsset function. This was not an issue before because we
always used file names without spaces, but starting with the dynamic
imports for disk definitions in 5899a7990a300fd4c71b5cb00949cd19633fa23e
we end up with "Sytem 1.0-<hash>.dsk.js" and similar files.