Stop using React.StrictMode around the entire app
Starting with React 18 it will unmount and remount every component
in dev mode (https://react.dev/blog/2022/03/08/react-18-upgrade-guide#updates-to-strict-mode).
This doesn't play well with the emulator (and makes recording of timings
more difficult). It doesn't add much value, so remove it from the root,
but keep it around the browser, which is the bulk of the rest of the
components are used.