Handle truncated disk images when mounting as CD-ROMs
The file size that we may be smaller than the size of the disk, since
empty data is omitted. While we could try to parse the full size out of
the partition map in the CD-ROM metadata handler, it's easier to just
allow chunks to be synthesized on the fly if reading or writing past
the end of the chunked file.

Updates #166