Make all/notable toggle selected state more obvious
The underline was chosen to avoid shifting the layout when toggling,
but it's not obvious that it indicates selected state (it could also
indicate that something is clickable).

Switch instead to using bold, and adding a hidden sizing element to
ensure that we always leave enough room for the bold version.