Add support for .bin/.cue CD-ROMs
For now we only handle MODE1/2352 images, and assume that the entire
.bin file is a single data track.

Works well enough for the sample files.

Also handles looking up CD-ROM URLs in the library when doing a custom
instance (so that we can save the metadata lookup).

Also adds a 2 second timeout for fetching a chunk (the Internet Archive
occasionally hangs, and better to fail and retry sooner).

Fixes #205
Updates #166