Add an debug_paused option
Audio playback requires a gesture to start, and we were immediately
starting the emulator. Since the only thing that DingusPPC can play
is the initial boot chime, it required clicking at just the right time
to be able to hear it. By making emulator start behind a click, we
can more deterministically trigger it.

Updates #219