Add wrapping of HFS disk images into full device images
Required for allowing DingusPPC to boot from disk images. Also allows Saved
HD to be exported to an .hda file that can be used with BlueSCSI and other
software that expects a full device image.

The base device image header was generated by Disk Jockey 2.5.2.2 (it's the
first 0xc000 or 48 KB of an empty device image - the HFS image can be appended
to it and a couple of fields be updated).

Updates #219
Updates #234