Add system CD-ROMs as drop-down options for custom instances
Makes it easy to start DingusPPC-based machines with CD-ROM images (since
the Beige G3 does not support hard drives yet).

Also cleans up CD-ROM library handling a bit - it can be obtained from
a global instead of being passed down as a prop (which was a remnant
from the days when it was only conditionally enabled).

Updates #219