Switch upload disks to being read in chunks
Similar to the work done for #56, we .iso's and other disk images that
are dragged in are also read in chunks. This reduces the memory
consuption fixes a pause when the disk is first mounted.

To keep things we can reuse the EmulatorWorkerChunkedDisk class, except
we load chunks slightly differently (HTTP range requests from the
File/Blob URL that is passed in from the UI process). This works as
expected in Blink and WebKit, but not in Gecko, so there we still
read the entire file (to minimize divergence in the two code paths,
we still treat it as one big chunk).