Preserve resource forks when dragging in folders
We can use the webkitGetAsEntry API to iterate over directories, and
in addition to uploading the base file, also upload its resource fork
by trying to read the ..namedfork/rsrc parallel file for it.

This only works in Chrome (Safari returns the File object, but trying to
read it fails, and Firefox doesn't even return the File object). It also
doesn't handle metadata like creator/type codes, so the files are not
directly usable even in Chrome. But it's sort of cool that it works at
all.

Updates #255