Redo System 1.0 image
Uses a combination of the System Disk and the Write/MacPaint disk from
https://www.earlymacintosh.org/disk_images.html#System_1 to have a
more authentic System 1.0 experience. MacPaint 1.0 is added to the
system disk, but MacWrite was already in the Infinite HD (MFS) image,
so it's left off (also because MacPaint needs some more free disk
space to actually run).

Also adds a pristine copy of the disk image as startup floppy, for
those that want to have the original experience.

Fixes #271