Allow custom instances to override the date
This gets converted into an offset from the current date, which is
then used to adjust the value that's returned by the emscripten_date_now
Wasm import.

To be able to hook the Emscripten imports we need to override
instantiateWasm and do our own instantiation, but that's actually
easier, because we can then more directly load the Wasm file (we
can fetch all dependencies as ArrayBuffers and transfer them to
the worker, instead of having to do some as blobs).

Fixes #199