Use view transitions when going from the browser to running Mac
We generate custom view transition names based on the disk that is
being run, so that we automatically map between the browser screen
frame and the one where the Mac is running.

Required a bit of rejiggering with lazy loading of the RunDefMac
component, but that should help with perceived performance in
general.