Checkpoint full library support
Adds a script for import a Macintosh Garden JSON dump and generating an
index file (small, loaded on the client) and details file (larger,
loaded in the worker).

Data files are stored as packed arrays to minimize download size (the
position in the array is a specific field of the item).

Adds a basic list and details view to the library drawer for showing
details about the item and triggering downloads.

Updates #337