Upgrade to Emscripten 3.1.74
It claims to have some small size improvements in the generated JS, though
the Wasm size is a bit bigger, so not a big change.

It does break the current approach of overriding the date in the emulator,
since emscripten_date_now is no longer exposed in the Wasm module imports
(I believe because emscripten-core/emscripten@c35d60a5ec1ef229195f6506e8e1b0a9659847a7
and/or emscripten-core/emscripten@c7f8724414c98f48a822a30790b0c12a20f39018
made it not cross library boundaries). We switch to overriding clock_time_get
as a fallback.

For now only DingusPPC is rebuilt with the new version of Emscripten.