Handle resolution changes from Mac OS X under PearPC
While it will honor a 640x480 size during early boot, it will try to
increase the resolution to 800x600 once the WindowServer loads. This
was leading to a JS exception because the SharedArrayBuffer used to
communicate between the emulator worker and UI threads was not increasing
in size. Pre-size it to the maximum likely size (1600x1200) so that we
don't need to worry about making it dynamically sizable.

Updates #357