Add NetNewsWire to the Mac OS X version of Infinite HD
1.0.2 is from https://download.macintoshgarden.org/apps/NetNewsWireLite1.0.2.dmg_.sit
but I had to unpack it manually into a .zip because modern hdiutil can no longer
mount that .dmg.

The 2.1.1 dmg works as is, but we do need to pipe in a `y` to the hdiutil command
to get it to accept the license agreement.

Also includes support for .dmg files inside compressed archives (which I ended
up not using).

Updates #357