Switch Mini vMac Mac II to debug builds
Appears to more reliably fix the stability issues with the System 6 Finder.

Also adds a Mac IIx machine option. Not materially different from the
Mac II (the emulated CPU is actually still a 68020), but do it for
completeness. Also make it have 32MB of RAM to experiment with that
now that it's possible (mihaip/minivmac@e1e7c4d77c0890425a4464d561baefa50174abb5).

The IIx ROM has checksum 97221136

Updates #405