Move disk chunks out of the public/ directory
They were moved to R2 for serving a while back, but we still had them
as static assets for local development. However, this was making
`vite build` take 1+ minute to complete (because it would copy the ~100K
chunks every time to to the build/ directory). Move them to a separate
directory (Images/build), and have the local middleware use that path.

For #390