Update to Emscripten 4.0.11
Required adding HEAPU8 to the exported symbols, but otherwise works well.

Drop the eval-based fallback mode for loading the generated JS in the
worker. It no longer works (due to more module syntax being added in the
generated JS), but it's also not needed anymore. It was added in
90c5abd51b6826b617d06399346f4f6033b08adb since Firefox's support for
modules in workers was very recent. It's now two years later, everyone
should have it.