Update to Emscripten 4.0.11
Required adding HEAPU8 to the exported symbols, but otherwise works well.

Drop the eval-based fallback mode for loading the generated JS in the
worker. It no longer works (due to more module syntax being added in the
generated JS), but it's also not needed anymore. It was added in
b221259595d323d2c45cd2175fdc7ce9287c3097 since Firefox's support for
modules in workers was very recent. It's now two years later, everyone
should have it.