Switch to wrangler.jsonc
Also rename the worker from `transparent-mac` (early name idea, when then focus was going to be more on exposing Mac internals) to `infinite-mac`.

Also drop the mac.persistent.info route - it was used in https://blog.persistent.info/2021/08/worker-loop.html but I updated that to point to infinitemac.org instead.