Make BlueSCSI support easier to use
Add the flag for enabling it to the custom instance configuration UI.

When enabled, mount a built-in "The Outside World" disk image that has a copy of the BlueSCSI SD Transfer app (the 1.1 beta that is more compatible) and a read me file explaining what to do.

To make the disk image friendlier, I set the "Invisible" bit on the debuglog.txt file. I also added a custom icon to the disk based on the AppleShare and WorldScript extension icons.

For #424